lambdaTest/requirements.txt
boto3
python-dotenv
pymongo[srv]